requirements.txt
ipykernel
python-dotenv
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
pandas
